Rename plotting to plotParser
